Fixed tests and moved formsteps to a folder
